pyproject.toml
[project]
name = "robo-advisor"
version = "0.1.0"
description = "A simple robo-advisor that uses the Alpaca API to get stock data and make recommendations."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.39.0",
    "dotenv>=0.9.9",
    "loguru>=0.7.3",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "yfinance>=0.2.54",
]
